Move ESQueryProfiler to more suitable location
